fix: fix preloads
