+++ packages/app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@deskofy/app",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "private": false,
   "description": "The Electron Framework",
   "author": "Deskofy",
