+++ packages/app/src/ipc/contexts/navigation.ts
@@ -2,18 +2,7 @@ import {
   BrowserWindow,
   BrowserWindowConstructorOptions,
   ipcMain,
-  ipcRenderer,
 } from 'electron';
-
-const navigationIpcContext = {
-  canGoBack: async (): Promise<any> =>
-    ipcRenderer.invoke('navigation:can-go-back'),
-  canGoForward: async (): Promise<any> =>
-    ipcRenderer.invoke('navigation:can-go-forward'),
-  goBack: (): void => ipcRenderer.send('navigation:go-back'),
-  goForward: (): void => ipcRenderer.send('navigation:go-forward'),
-  cloneWindow: (): void => ipcRenderer.send('navigation:clone-window'),
-};
 
 const registerNavigationIpcContext = (
   browserWindowData: BrowserWindowConstructorOptions | undefined,
@@ -63,4 +52,4 @@ const registerNavigationIpcContext = (
   });
 };
 
-export { navigationIpcContext, registerNavigationIpcContext };
+export { registerNavigationIpcContext };
