+++ packages/app/src/ipc/preload/version.ts
